requirements.txt
apache-airflow[gcp]
google-cloud-bigquery
pandas
scikit-learn
jupyter
notebook
nbconvert
ipykernel
joblib==1.4.2
matplotlib
seaborn
xgboost
cloudpickle 
papermill 
